requirements.txt
python-dotenv
Flask
google-generativeai